COMP: add casting to satisfy pow
svn-url: http://viewvc.slicer.org/viewvc.cgi/Slicer4?view=revision&revision=5285
git-svn-id: http://svn.slicer.org/Slicer4/trunk@5285 3bd1e089-480b-0410-8dfb-8563597acbee

+++ Modules/EMSegment/Registration/vtkBSplineRegistrator.cxx
@@ -427,7 +427,8 @@ RegisterImagesInternal3()
 
   int initialNumberOfLevels = 3;
   int initialShrinkFactor = 
-    static_cast<int>(std::pow(2,static_cast<float>(initialNumberOfLevels)));
+    static_cast<int>(std::pow(static_cast<double>(2.0),
+                              static_cast<double>(initialNumberOfLevels)));
 
   fixedImagePyramid->SetNumberOfLevels(initialNumberOfLevels);
   fixedImagePyramid->SetStartingShrinkFactors(initialShrinkFactor);
